Hole detector architecture shift
Until now, the synchronizer node published two messages:
One to a topic where the depth node subscribed to, that contained
the point cloud, and one to a topic where the RGB node subscribed to,
that contained the point-cloud-extracted RGB image. The Depth node
extracted the depth image from the point cloud, and published the
depth image, the point cloud and the candidate holes found to the hole
fusion node. The point cloud, though, was not needed by the depth node.

Now, the synchronizer node publishes the depth image to a topic where
the depth node is subscribed to, the rgb image to a topic where the rgb
node is subscribed to and the point cloud directly to the hole fusion
node.

So, this commit closes #15.

Furthermore, the point cloud on which the filter of plane constiturion
was based, was the original point cloud, that is, containing noise.
Now, the hole fusion node takes the point cloud and interpolates it at
the same time when the depth and rgb nodes are processing the images
sent by the synchronizer node, in order to save on execution time.

So, this commit resolves #17.

	modified:   pandora_vision_hole_detector/CMakeLists.txt
	modified:   pandora_vision_hole_detector/include/depth_node/depth.h
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/hole_fusion.h
	modified:   pandora_vision_hole_detector/include/rgb_dummy_node/rgb_dummy.h
	modified:   pandora_vision_hole_detector/include/rgb_node/hole_detection.h
	modified:   pandora_vision_hole_detector/include/synchronizer_node/rgb_depth_synchronizer.h
	modified:   pandora_vision_hole_detector/include/utils/message_conversions.h
	modified:   pandora_vision_hole_detector/src/depth_node/depth.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_fusion.cpp
	modified:   pandora_vision_hole_detector/src/rgb_dummy_node/rgb_dummy.cpp
	modified:   pandora_vision_hole_detector/src/rgb_node/hole_detection.cpp
	modified:   pandora_vision_hole_detector/src/synchronizer_node/rgb_depth_synchronizer.cpp
	modified:   pandora_vision_hole_detector/src/utils/message_conversions.cpp
	modified:   vision_communications/CMakeLists.txt
	renamed:    vision_communications/msg/RgbCandidateHolesVectorMsg.msg -> vision_communications/msg/CandidateHolesVectorMsg.msg
	deleted:    vision_communications/msg/DepthCandidateHolesVectorMsg.msg