Major changes
Hole filters, either depth or rgb ones, up until now constructed
tools needed for validating holes on their own, like inflating
rectangles, or using the brushfire algorithm to find the intermediate
points between a hole's outline and its inflated bounding rectangle.
These tools are finite and there wasn't a one-to-one correspondence
between a filter and a tool.
Hence, to save on execution resources, the construction of these tools
is now made once, centrally, and distributed to each filter that is in
need of said tools.

Second, the raycast algorithm, that was used to identify a hole's
outline, up until now returned a non-coherent set of points.
Now it returns the exact outline of a hole, that is, a set of points
that is a closed line.

Third, the combination of the above two changes made it possible to
ditch the cv::pointPolygonTest and replace it with the brushfire
algorithm, which is much faster.
Intermediate points, holes' masks etc are now produced via
the brushfire algorithm because it allows for the filling of closed
shapes.

	modified:   pandora_vision_hole_detector/cfg/hole_fusion_cfg.cfg
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/depth_filters.h
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/hole_fusion.h
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/rgb_filters.h
	modified:   pandora_vision_hole_detector/include/utils/blob_detection.h
	modified:   pandora_vision_hole_detector/include/utils/parameters.h
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/depth_filters.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/generic_filters.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_fusion.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/rgb_filters.cpp
	modified:   pandora_vision_hole_detector/src/utils/blob_detection.cpp
	modified:   pandora_vision_hole_detector/src/utils/holes_conveyor.cpp
	modified:   pandora_vision_hole_detector/src/utils/parameters.cpp