Fixes #16
The problem with #16 was that then the same hole with a slightly
different outline was presented by both RGB and depth sources, the
cv::pointPolygonTest test regarded points that coinsided in space and
therefore should give a result of 0, outside of the outline, even though
the given point was actually ON the edge.

This kind of anomally meant that the solution should employ the
brushfire algorithm, given that the outline points of a hole is now a
coherent shape.

Also, when considering of connecting distant holes, a new check is
implemented. Besides the proximity check, a second distance check was
added and pertains to the distance between the farthest two points of
the two to-be-connected holes.

Lastly, concerning #23, the cv::pointPolygonTest method is now called
only on contours that are not closed shapes; this for sure works at
least

	modified:   pandora_vision_hole_detector/cfg/hole_fusion_cfg.cfg
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/hole_merger.h
	modified:   pandora_vision_hole_detector/include/utils/blob_detection.h
	modified:   pandora_vision_hole_detector/include/utils/holes_conveyor.h
	modified:   pandora_vision_hole_detector/include/utils/parameters.h
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_fusion.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_merger.cpp
	modified:   pandora_vision_hole_detector/src/utils/blob_detection.cpp
	modified:   pandora_vision_hole_detector/src/utils/holes_conveyor.cpp
	modified:   pandora_vision_hole_detector/src/utils/parameters.cpp