Significant changes in edge detection / denoising
One issue so far was that holes' outlines where de facto approximations
of the real outlines. That was located in the edge detection and edge
denoising methodology followed. The initial dilation performed,
after the detection of edges on the depth or rgb images,
disfigured the initial outlines and this error was propagated throughout
the rest of the package's processes.

As of now, this initial dilation of the image's edges is canceled, and
so is the thinning made afterwards (the edges contamination process
functions flawlessly, although the thinning process could be enabled in
order to ease it).

The second major change is that, as of now, the Canny edge detector is
used, instead of the Sobel one. Also, the Scharr one is not that bad
either; we should consider it, especially in the edge detection on the
depth image.

	modified:   pandora_vision_hole_detector/cfg/hole_fusion_cfg.cfg
	modified:   pandora_vision_hole_detector/cfg/rgb_cfg.cfg
	modified:   pandora_vision_hole_detector/include/utils/parameters.h
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_fusion.cpp
	modified:   pandora_vision_hole_detector/src/rgb_node/rgb.cpp
	modified:   pandora_vision_hole_detector/src/utils/edge_detection.cpp
	modified:   pandora_vision_hole_detector/src/utils/parameters.cpp