Introduces new getShapesClearBorder method
While the getShapesClearBorder method identifies all closed regions,
even those inside other closed regions, if the amount of noise inside a
closed region is excessive, the method may fail to identify the
outermost closed region as a blob.

In order to achieve the above result, a getShapesClearBorderSimple
method is introduced, which disregards the clutter inside the outermost
closed shape and considers its outline as the only and valid outline of
the shape.

The downside is that if, by any chance, as mentioned in #14, there are
closed shapes inside other ones, that should still be considered valid,
the use of this method fails to do so.

	modified:   pandora_vision_hole_detector/cfg/depth_cfg.cfg
	modified:   pandora_vision_hole_detector/cfg/hole_fusion_cfg.cfg
	modified:   pandora_vision_hole_detector/cfg/rgb_cfg.cfg
	modified:   pandora_vision_hole_detector/include/utils/edge_detection.h
	modified:   pandora_vision_hole_detector/src/utils/edge_detection.cpp