Adds improvement on the checkHolesTextureBackProject filter
A new dilation method was added, one that takes account not only binary
images, but dilates a zero pixel giving it the non-zero value (not
necessarily 255) of one of each immediate neighbors.

The backproject image obtained in the checkHolesTextureBackProject
filter is now dilated using this operator to obtain a more accurate
depiction of what should be the backproject (due to its often sparce
nature).

	modified:   pandora_vision_hole_detector/CMakeLists.txt
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/rgb_filters.h
	modified:   pandora_vision_hole_detector/include/utils/morphological_operators.h
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/rgb_filters.cpp
	modified:   pandora_vision_hole_detector/src/rgb_node/hole_detector.cpp
	modified:   pandora_vision_hole_detector/src/utils/morphological_operators.cpp