Revolutionizes RGB edges extraction
The edges extraction of the synchronized RGB image is now made using two
fashions:

(a) The first choice is to segment the image, posterize it, convert it
to grayscale and then find its edges. This process implicates a number
of parameters that need to be set accurately so that the posterization
process does not destroy the shapes of the image.

(b) The second choice involves the backprojection of the segmented RGB
image, and the use of the watershed algorithm in order to find whole
regions, even if the backprojection image's integrity is compromised,
whose histogram matches the model input histogram. The edges of the RGB
image are then extracted immediately from the output of the watershed
algorithm.

	modified:   pandora_vision_hole_detector/cfg/rgb_cfg.cfg
	modified:   pandora_vision_hole_detector/include/rgb_node/hole_detector.h
	modified:   pandora_vision_hole_detector/include/utils/parameters.h
	modified:   pandora_vision_hole_detector/src/rgb_node/hole_detector.cpp
	modified:   pandora_vision_hole_detector/src/rgb_node/rgb.cpp
	modified:   pandora_vision_hole_detector/src/utils/parameters.cpp