Package pandora_vision_hole_detector works with simulator
(a) Depending on the source of the point cloud, real time via a physical
depth sensor or by simulation, its parameters are altered in order for
the package to function properly.

(b) The parameters pertaining to the size of the point cloud and its
data are now passed to the package via the nodehandle, having as a start
point the launch file used to commence the execution of the package.

(c) The assimilation process ceases to have elevated privileges in case
of filtering: all merge operations are checked before actually merging
occurs.

	modified:   pandora_vision_hole_detector/include/utils/parameters.h
	modified:   pandora_vision_hole_detector/launch/pandora_vision_hole_detector.launch
	deleted:    pandora_vision_hole_detector/launch/pandora_vision_hole_detector_dry.launch
	renamed:    pandora_vision_hole_detector/launch/pandora_vision_hole_detector.launch -> pandora_vision_hole_detector/launch/pandora_vision_hole_detector_standalone.launch
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_merger.cpp
	modified:   pandora_vision_hole_detector/src/synchronizer_node/rgb_depth_synchronizer.cpp
	modified:   pandora_vision_hole_detector/src/utils/message_conversions.cpp
	modified:   pandora_vision_hole_detector/src/utils/parameters.cpp