Closes #48 and fixes #49
Also, since the RgbFilters::checkHolesTextureBackProject method now uses
a watershed approach, together with
EdgeDetection::produceEdgesViaBackprojection, the core of these two
functionalities has now moved to a separate function:
EdgeDetection::watershedViaBackprojection. It takes as input a RGB image
and a backprojection and outputs a homogenous backprojection image in
the former case and an image containing the edges of the RGB image in
the latter.

	modified:   pandora_vision_hole_detector/cfg/depth_cfg.cfg
	modified:   pandora_vision_hole_detector/cfg/hole_fusion_cfg.cfg
	modified:   pandora_vision_hole_detector/cfg/rgb_cfg.cfg
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/rgb_filters.h
	modified:   pandora_vision_hole_detector/include/utils/edge_detection.h
	modified:   pandora_vision_hole_detector/include/utils/parameters.h
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_fusion.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/rgb_filters.cpp
	modified:   pandora_vision_hole_detector/src/rgb_node/rgb.cpp
	modified:   pandora_vision_hole_detector/src/utils/edge_detection.cpp
	modified:   pandora_vision_hole_detector/src/utils/parameters.cpp