Adds control parameter in DepthFilters::checkHolesDepthDiff
Until now, the probability assignment to a hole's validity in the depth
difference filter used a normal distribution around the value of an
expected difference in distance from the depth sensor, between the
keypoint and the four vertices of the hole's bounding box.
Because it is not certain that items inside the vicinity of the hole
will be deep inside it
(e.g.  http://gasi.sorocaba.unesp.br/wordpress/wp-content/uploads/2013/11/9157813829_8bb119311a_b.jpg),
a parameter was added to control the method by which the validity
probability of a hole is determined. Now, it is possible to obtain a
binary probability in the case that a positive difference in distances
is detected, instead of a continuous function-based one.

	modified:   pandora_vision_hole_detector/cfg/hole_fusion_cfg.cfg
	modified:   pandora_vision_hole_detector/include/utils/parameters.h
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/depth_filters.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_fusion.cpp
	modified:   pandora_vision_hole_detector/src/utils/parameters.cpp
	modified:   pandora_vision_hole_detector/test/unit/hole_fusion_node/depth_filters_test.cpp
	modified:   pandora_vision_hole_detector/test/unit/utils/edge_detection_test.cpp