Closes #51
The RGB and Depth filters execution was moved from their respective
domain and moved into a new class which applies each one of them, with
an order explicitly specified for each one, within the overall order of
execution, whereas, until now, the order of each filter was specified in
its respective domain (RGB or Depth).

In order to facilitate this process, because depth analysis, in general,
is not always possible, a new system of execution order was introduced
that pertains only to when RGB filters can be applied.

Each order must be specified with a value within the set of {1, .. N}
with distinct increments of one, or else segmentation fault WILL occur.

	modified:   pandora_vision_hole_detector/CMakeLists.txt
	modified:   pandora_vision_hole_detector/cfg/hole_fusion_cfg.cfg
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/depth_filters.h
	new file:   pandora_vision_hole_detector/include/hole_fusion_node/filters.h
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/hole_fusion.h
	modified:   pandora_vision_hole_detector/include/hole_fusion_node/rgb_filters.h
	modified:   pandora_vision_hole_detector/include/utils/parameters.h
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/CMakeLists.txt
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/depth_filters.cpp
	new file:   pandora_vision_hole_detector/src/hole_fusion_node/filters.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_fusion.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_merger.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/rgb_filters.cpp
	modified:   pandora_vision_hole_detector/src/utils/parameters.cpp