Fixes tests due to 91a12ee4553784e8c7cb69a6a8f8f27905acf66b changes.
Fixes indices bug in the rgb filters.
The holes mask vector has a size equal to the number of candidate holes,
however, the intermediate points vector shall have only as maximum this
number, and will not always be equal to it: the inflated rectangle of
some holes might exceed the image's bounds.

	modified:   pandora_vision_hole_detector/src/hole_fusion_node/rgb_filters.cpp
	modified:   pandora_vision_hole_detector/test/unit/hole_fusion_node/filters_test.cpp
	modified:   pandora_vision_hole_detector/test/unit/hole_fusion_node/rgb_filters_test.cpp
	modified:   pandora_vision_hole_detector/test/unit/rgb_node/hole_detector_test.cpp
	modified:   pandora_vision_hole_detector/test/unit/utils/edge_detection_test.cpp
	deleted:    pandora_vision_hole_detector/walls/texture_1/0.png
	deleted:    pandora_vision_hole_detector/walls/texture_2/0.png